+++ src/Plutarch/Extra/StateThread.hs
@@ -1,5 +1,6 @@
 module Plutarch.Extra.StateThread (
   withStateThread,
+  pwithStateThread,
 ) where
 
 import Plutarch.Api.V1 (PCurrencySymbol, PValue)
@@ -7,37 +8,48 @@ import Plutarch.Api.V1.Value (pvalueOf)
 import Plutarch.Api.V2 (
   AmountGuarantees,
   KeyGuarantees,
-  PScriptContext,
+  PMintingPolicy,
   PScriptPurpose (PMinting),
   PTxInInfo,
+  PTxOutRef,
  )
 import Plutarch.Extra.Field (pletAll)
-import PlutusLedgerApi.V2 (TxOutRef)
 
 {- | Adds a state thread to a minting policy.
+ Parameterized at the Haskell level.
 
- @since 3.9.2
+ @since 3.11.0
 -}
 withStateThread ::
   forall (s :: S).
-  -- | Initial spend
-  TxOutRef ->
   -- | Minting policy to wrap
-  Term s (PData :--> PScriptContext :--> POpaque) ->
-  Term s (PData :--> PScriptContext :--> POpaque)
-withStateThread ref mp = plam $ \red ctx -> pletAll ctx $ \ctx' ->
+  Term s PMintingPolicy ->
+  -- | Initial spend
+  Term s PTxOutRef ->
+  Term s PMintingPolicy
+withStateThread mp ref = plam $ \red ctx -> pletAll ctx $ \ctx' ->
   pletFields @'["inputs", "mint"] (getField @"txInfo" ctx') $ \txInfo ->
     pmatch (getField @"purpose" ctx') $ \case
       PMinting thisPolicy ->
         pif
           (uniqueStateTokenMint (pfield @"_0" # thisPolicy) . getField @"mint" $ txInfo)
           ( pif
-              (pany # hasUniqueInput ref # getField @"inputs" txInfo)
+              (pany # (hasUniqueInput # ref) # getField @"inputs" txInfo)
               (mp # red # ctx)
               (ptraceError "stateThread: Unique input not found")
           )
           (ptraceError "stateThread: Not minting a unique state token")
       _ -> ptraceError "stateThread: Not a minting script purpose"
+
+{- | Adds a state thread to a minting policy.
+ Parameterized at the Plutarch level
+
+ @since 3.11.0
+-}
+pwithStateThread ::
+  forall (s :: S).
+  Term s (PMintingPolicy :--> PTxOutRef :--> PMintingPolicy)
+pwithStateThread = plam withStateThread
 
 -- Helpers
 
@@ -51,7 +63,6 @@ uniqueStateTokenMint thisPolicy mint =
 
 hasUniqueInput ::
   forall (s :: S).
-  TxOutRef ->
-  Term s (PTxInInfo :--> PBool)
-hasUniqueInput ref =
-  plam $ \txInInfo -> pconstantData ref #== (pfield @"outRef" # txInInfo)
+  Term s (PTxOutRef :--> PTxInInfo :--> PBool)
+hasUniqueInput =
+  plam $ \ref txInInfo -> ref #== (pfield @"outRef" # txInInfo)